lambda/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.12.1 ; python_version >= "3.11" and python_version < "4.0"
ask-sdk-core==1.19.0 ; python_version >= "3.11" and python_version < "4.0"
ask-sdk-model==1.82.0 ; python_version >= "3.11" and python_version < "4.0"
ask-sdk-runtime==1.19.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2026.2.25 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.6 ; python_version >= "3.11" and python_version < "4.0"
cryptography==46.0.5 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.49.1 ; python_version >= "3.11" and python_version < "4.0"
google-genai==1.68.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
pycparser==3.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.2 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
tenacity==9.1.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
websockets==16.0 ; python_version >= "3.11" and python_version < "4.0"
